Move document copy-paste tests to own module.
